move Assets icons to images
