+++ src/Assets/images/scrap-hovered.png
